fix: memoizzazione in ricalcolo prezzi dei gemelli
- quando si salva una placca che deve aggiornare anche i costi delle
placche "gemelle", diversi dati che non cambiano vengono ora memoizzati
- sweep verticale unico per tutti i gemelli + flag skipPrewarm a valle
- memo per request nel pricing: zona, productItem, productMap, metadata,
ricerche bundle/prodotto
- update: salta la riscrittura frutti/posizioni/product item se l'hash
non cambia

+++ com/apirone/core/model/service/QuotationItemService.cfc
@@ -173,20 +173,33 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 	}
 
 	public String function update( required com.apirone.core.model.bean.QuotationItem quotationItem ){
-
+		// Se l'impronta dell'item (calcolata sui dati in memoria) coincide con quella salvata
+		// nel bean (hash letta dal DB durante il build), la configurazione non è cambiata:
+		// l'hash esclude prezzo e quantità ma include frutti, posizioni, note e tappi
+		// ricalcolati, quindi salta la riscrittura completa di frutti/posizioni/product item
+		// (tipico dell'aggiornaPrezzo dei gemelli, che cambia solo il prezzo).
 		// Il confronto fra i frutti serve solo per gli ID: legge gli id esistenti
 		// con una sola query invece di ricaricare e ricostruire l'intero item con
 		// getMany() (cascata N+1 di build completi)
+		// Se l'hash salvato non esiste (item nuovo o mai persistito) il confronto non scatta
+		// e la riscrittura completa dei frutti avviene come prima.
 		var oldFruitIds = [];
+		var fruitIdsToDeleted = [];
+		var doSkipFruits = false;
+		var newHash = NullValue();
 		if ( IsInstanceOf( arguments.quotationItem, "com.apirone.core.model.bean.QuotationItemPlate" ) ) {
+			newHash = getProductHashService().createHash( arguments.quotationItem.getId(), arguments.quotationItem );
+			var storedHash = arguments.quotationItem.getHash();
+			doSkipFruits = !IsNull( storedHash ) && !IsNull( newHash ) && storedHash == newHash;
+		}
+		if ( IsInstanceOf( arguments.quotationItem, "com.apirone.core.model.bean.QuotationItemPlate" ) && !doSkipFruits ) {
 			var fruitRows = getQuotationItemFruitService().getDao().findByQuotationItemIds( [ arguments.quotationItem.getId() ] );
 			for ( var fruitRow in fruitRows ) {
 				oldFruitIds.append( fruitRow.quotation_item_fruit_id );
 			}
 		}
 
-		if ( IsInstanceOf( arguments.quotationItem, "com.apirone.core.model.bean.QuotationItemPlate" ) ) {
-			var fruitIdsToDeleted = [];
+		if ( IsInstanceOf( arguments.quotationItem, "com.apirone.core.model.bean.QuotationItemPlate" ) && !doSkipFruits ) {
 
 			// Raccoglie gli id dei frutti nuovi valorizzati
 			// (i tappi ricalcolati hanno id vuoto)
@@ -219,7 +232,7 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 
 			getDao().update( arguments.quotationItem );
 
-			if ( IsInstanceOf( arguments.quotationItem, "com.apirone.core.model.bean.QuotationItemPlate" ) ) {
+			if ( IsInstanceOf( arguments.quotationItem, "com.apirone.core.model.bean.QuotationItemPlate" ) && !doSkipFruits ) {
 				for( var thisFruitId in fruitIdsToDeleted ) {
 					getQuotationItemFruitService().delete( thisFruitId );
 				}
@@ -249,7 +262,7 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 				// Passa il bean già caricato: l'hash non contiene id di riga, quindi il
 				// bean in memoria (già sincronizzato con i frutti ricomputati) produce
 				// lo stesso hash di un reload completo
-				var hash = getProductHashService().createHash( arguments.quotationItem.getId(), arguments.quotationItem );
+				var hash = IsNull( newHash ) ? getProductHashService().createHash( arguments.quotationItem.getId(), arguments.quotationItem ) : newHash;
 				if ( !IsNull( hash ) ) {
 					updateHash( arguments.quotationItem.getId(), hash );
 				}
@@ -945,10 +958,17 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 	public com.apirone.core.model.bean.QuotationItemPrice function getPlatePricing(
 		required Struct data,
 		preloadedQuotation = javacast( "null", "" ),
-		preloadedQuotationItem = javacast( "null", "" )
+		preloadedQuotationItem = javacast( "null", "" ),
+		skipPrewarm = javacast( "null", "" )
 	){
 
 		var json = arguments.data;
+
+		// Lettura difensiva di skipPrewarm: quando il chiamante non lo passa, Lucee NON binda
+		// un argomento posizionale con valore null (la chiave resterebbe numerica nell'arguments
+		// scope) e un accesso diretto lancerebbe "key doesn't exist". Da qui in poi viaggia
+		// solo un booleano, che si binda sempre.
+		var doSkipPrewarm = StructKeyExists( arguments, "skipPrewarm" ) && !IsNull( arguments.skipPrewarm ) && arguments.skipPrewarm;
 
 		var pricing = super.bean( "QuotationItemPrice" );
 		var method  = super.bean( "PriceMethod" );
@@ -1019,7 +1039,21 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 			ArrayAppend( productIdsToPreload, variables.plugProductIdList, true );
 		}
 
-		var productMap = getProductService().getMany( productIdsToPreload );
+		// Memo per request: gli stessi prodotti (placca + frutti + tappi) tornano in tutti gli
+		// articoli gemelli; i bean prodotto sono usati in sola lettura nel pricing
+		var pmpKey = ArrayToList( productIdsToPreload );
+		if ( !StructKeyExists( request, "_pricingProductMapMemo" ) ) {
+			request._pricingProductMapMemo = {};
+		}
+		var productMap = NullValue();
+		if ( ArrayLen( productIdsToPreload ) && StructKeyExists( request._pricingProductMapMemo, pmpKey ) ) {
+			productMap = request._pricingProductMapMemo[ pmpKey ];
+		} else if ( ArrayLen( productIdsToPreload ) ) {
+			productMap = getProductService().getMany( productIdsToPreload );
+			request._pricingProductMapMemo[ pmpKey ] = productMap;
+		} else {
+			productMap = {};
+		}
 
 		// Prewarm verticale: una sola coppia di query listin/artico per TUTTA la placca
 		// (componenti degli item selezionati di placca e frutti + componenti dei prodotti
@@ -1035,12 +1069,20 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 			}
 		}
 
-		getComponentService().prewarmPricingComponents(
-			productItemIds = allProductItemIds,
-			productIds     = productIdsToPreload,
-			lineId         = productMap[ product.id ].getLine().getId(),
-			modelId        = productMap[ product.id ].getModel().getId()
-		);
+		// skipPrewarm arriva solo dal ricalcolo a catena dei prezzi (l'aggregatore degli
+		// articoli gemelli, vedi aggiornaPrezzoAltriArticoliByQuotationIdLineIdFinishId, che
+		// esegue UNO sweep verticale per tutte le placche prima del loop). In quel contesto
+		// le memo ERP contengono già i costi di tutti i componenti: il prewarm locale sarebbe
+		// uno spreco (diverse letture apirone per niente). Dal salvataggio normale (client)
+		// il flag non arriva e il prewarm locale gira come sempre.
+		if ( !doSkipPrewarm ) {
+			getComponentService().prewarmPricingComponents(
+				productItemIds = allProductItemIds,
+				productIds     = productIdsToPreload,
+				lineId         = productMap[ product.id ].getLine().getId(),
+				modelId        = productMap[ product.id ].getModel().getId()
+			);
+		}
 
 		var platePrice = calculator.calculate(
 			product.id,
@@ -1051,7 +1093,8 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 			0,
 			quotation,
 			quotationItem,
-			StructKeyExists( productMap, product.id ) ? productMap[ product.id ] : NullValue()
+			StructKeyExists( productMap, product.id ) ? productMap[ product.id ] : NullValue(),
+			doSkipPrewarm
 		);
 
 		var line = super.bean( "QuotationItemPriceLine" );
@@ -1088,7 +1131,8 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 				0,
 				quotation,
 				quotationItem,
-				StructKeyExists( productMap, fruit.fruit.id ) ? productMap[ fruit.fruit.id ] : NullValue()
+				StructKeyExists( productMap, fruit.fruit.id ) ? productMap[ fruit.fruit.id ] : NullValue(),
+				doSkipPrewarm
 			);
 
 			line.setName( "#fruit.fruit?.name#" );
@@ -1113,7 +1157,8 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 					0,
 					quotation,
 					quotationItem,
-					plugProduct
+					plugProduct,
+					doSkipPrewarm
 				);
 				var plugLine = super.bean( "QuotationItemPriceLine" );
 				plugLine.setName( plug.type == "tappo" ? "Tappo" : "Mezzo tappo" );
@@ -1542,6 +1587,78 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 		}
 		var itemMap = ArrayLen( itemIds ) ? getMany( itemIds ) : {};
 
+		// Prewarm verticale per TUTTI i gemelli in una passata: raccoglie item, prodotti
+		// (placca + frutti + tappi) e modelli distinti dai bean già caricati e fa una sola
+		// coppia di query listin/artico per l'intero batch. Senza questo sweep ogni gemella
+		// farebbe le proprie letture ERP (decine di round trip su preventivi con molte placche).
+		// I prewarm interni dei singoli aggiornaPrezzo vengono poi saltati (skipPrewarm),
+		// perché tutto il necessario è già nelle memo.
+		var hoistItemIds    = [];
+		var hoistProductIds = [];
+		var hoistModelIds   = {};
+		var hoistSeen       = {};
+
+		for ( var sibKey in itemMap ) {
+			var sib = itemMap[ sibKey ];
+
+			if ( !IsNull( sib.getProduct() ) ) {
+				var sibProductId = sib.getProduct().getId();
+				if ( !StructKeyExists( hoistSeen, "p" & sibProductId ) ) {
+					hoistSeen[ "p" & sibProductId ] = true;
+					hoistProductIds.append( sibProductId );
+				}
+				if ( !IsNull( sib.getProduct().getModel() ) ) {
+					hoistModelIds[ sib.getProduct().getModel().getId() ] = true;
+				}
+			}
+
+			if ( !IsNull( sib.getItems() ) ) {
+				for ( var sibItem in sib.getItems() ) {
+					var sibItemPiId = sibItem.getProductItem().getId();
+					if ( !StructKeyExists( hoistSeen, "i" & sibItemPiId ) ) {
+						hoistSeen[ "i" & sibItemPiId ] = true;
+						hoistItemIds.append( sibItemPiId );
+					}
+				}
+			}
+
+			if ( !IsNull( sib.getFruits() ) ) {
+				for ( var sibFruit in sib.getFruits() ) {
+					if ( !IsNull( sibFruit.getFruit() ) ) {
+						var sibFruitProductId = sibFruit.getFruit().getId();
+						if ( !StructKeyExists( hoistSeen, "p" & sibFruitProductId ) ) {
+							hoistSeen[ "p" & sibFruitProductId ] = true;
+							hoistProductIds.append( sibFruitProductId );
+						}
+					}
+					if ( !IsNull( sibFruit.getItems() ) ) {
+						for ( var sibFruitItem in sibFruit.getItems() ) {
+							var sibFruitPiId = sibFruitItem.getProductItem().getId();
+							if ( !StructKeyExists( hoistSeen, "i" & sibFruitPiId ) ) {
+								hoistSeen[ "i" & sibFruitPiId ] = true;
+								hoistItemIds.append( sibFruitPiId );
+							}
+						}
+					}
+				}
+			}
+		}
+
+		// tappi (viti a vista no): getPlatePricing precarica anche i plug products
+		for ( var plugProductId in variables.plugProductIdList ) {
+			if ( !StructKeyExists( hoistSeen, "p" & plugProductId ) ) {
+				hoistSeen[ "p" & plugProductId ] = true;
+				hoistProductIds.append( plugProductId );
+			}
+		}
+
+		getComponentService().prewarmPricingComponents(
+			productItemIds = hoistItemIds,
+			productIds     = hoistProductIds,
+			lineId         = arguments.lineId,
+			modelIds       = StructKeyArray( hoistModelIds )
+		);
+
 		// La Quotation è la stessa per tutte le righe: viene estratta una sola volta dai bean
 		// già caricati in batch e passata a ogni aggiornaPrezzo (evita il rebuild per item, N+1).
 		var sharedQuotation = NullValue();
@@ -1567,7 +1684,7 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 				continue;
 			}
 
-			aggiornaPrezzo( quotationItem, sharedQuotation );
+			aggiornaPrezzo( quotationItem, sharedQuotation, true );
 		}
 	}
 
@@ -1625,8 +1742,11 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 		aggiornaPrezzoAltriArticoli...) di passare la Quotation già caricata in
 		batch, evitando un refetch completo della Quotation per ogni riga (N+1).
 	*/
-	public function aggiornaPrezzo( required quotationItem, preloadedQuotation = javacast( "null", "" ) )
+	public function aggiornaPrezzo( required quotationItem, preloadedQuotation = javacast( "null", "" ), skipPrewarm = javacast( "null", "" ) )
 	{
+		// skipPrewarm letto con StructKeyExists: un null passato posizionalmente non viene bindato
+		var doSkipPrewarm = StructKeyExists( arguments, "skipPrewarm" ) && !IsNull( arguments.skipPrewarm ) && arguments.skipPrewarm;
+
 		if (!isNull(quotationItem.getArticle())) {
 			return false;
 		}
@@ -1731,7 +1851,8 @@ component extends="com.apirone.core.model.service.AbsService" accessors="true" {
 			allFruits.append( realFruits, true );
 			allFruits.append( plugBeans, true );
 			quotationItem.setFruits( allFruits );
-			var price = getPlatePricing( json, preloadedQuotation, quotationItem )
+			// doSkipPrewarm: l'eventuale sweep di massa è già stato fatto dall'aggregatore
+			var price = getPlatePricing( json, preloadedQuotation, quotationItem, doSkipPrewarm )
 		}
 
 		if (IsInstanceOf(quotationItem, "com.apirone.core.model.bean.QuotationItemSignage")) {
